move generated catalog assertion logic into super class
